refactor: :recycle: refactor to make integration tests possible
