+++ src/lib.rs
@@ -1,0 +1,22 @@
+use actix_web::{web, App, HttpRequest, HttpResponse, HttpServer, Responder};
+
+async fn greet(req: HttpRequest) -> impl Responder {
+    let name = req.match_info().get("name").unwrap_or("World");
+    format!("Hello {}!", &name)
+}
+
+async fn health_check() -> impl Responder {
+    HttpResponse::Ok().finish()
+}
+
+pub async fn run() -> Result<(), std::io::Error> {
+    HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/", web::get().to(greet))
+            .route("/{name}", web::get().to(greet))
+    })
+    .bind("127.0.0.1:8000")?
+    .run()
+    .await
+}
